requirements.txt
# Graph RAG System Requirements
pandas>=2.3.3
openpyxl>=3.1.5
sentence-transformers>=5.1.1
faiss-cpu>=1.12.0
numpy>=1.26.0
anthropic>=0.40.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pymupdf>=1.26.4
networkx>=3.0
matplotlib>=3.5.0
neo4j>=5.15.0
